Revert "LRP-4526 Cast to int to keep the same logic as before"
This reverts commit f1200cf0952d6ca40fbc3667c447f882b783e7f8.

+++ tools/release/ext/ext-impl/src/com/liferay/portal/license/LicenseManager.java
@@ -942,7 +942,7 @@ public class LicenseManager {
 				pos = 0;
 			}
 			else {
-				pos += (int)c;
+				pos += c;
 			}
 		}
 	}
